Unified logic message cells logic. Added sending/failed states to image message
